# perf(parser): remove `-> Result<T>` from all parsing methods