--- a/crates/oxc_parser/src/js/binding.rs
+++ b/crates/oxc_parser/src/js/binding.rs
@@ -1,5 +1,4 @@
 use oxc_ast::{NONE, ast::*};
-use oxc_diagnostics::Result;
 use oxc_span::GetSpan;
 
 use crate::{Context, ParserImpl, diagnostics, lexer::Kind};
@@ -8,28 +7,25 @@ impl<'a> ParserImpl<'a> {
     /// `BindingElement`
     ///     `SingleNameBinding`
     ///     `BindingPattern`[?Yield, ?Await] `Initializer`[+In, ?Yield, ?Await]opt
-    pub(super) fn parse_binding_pattern_with_initializer(&mut self) -> Result<BindingPattern<'a>> {
+    pub(super) fn parse_binding_pattern_with_initializer(&mut self) -> BindingPattern<'a> {
         let span = self.start_span();
-        let pattern = self.parse_binding_pattern(true)?;
+        let pattern = self.parse_binding_pattern(true);
         self.context(Context::In, Context::empty(), |p| p.parse_initializer(span, pattern))
     }
 
-    pub(super) fn parse_binding_pattern(
-        &mut self,
-        allow_question: bool,
-    ) -> Result<BindingPattern<'a>> {
-        let mut kind = self.parse_binding_pattern_kind()?;
+    pub(super) fn parse_binding_pattern(&mut self, allow_question: bool) -> BindingPattern<'a> {
+        let mut kind = self.parse_binding_pattern_kind();
         let optional = if allow_question && self.is_ts { self.eat(Kind::Question) } else { false };
-        let type_annotation = self.parse_ts_type_annotation()?;
+        let type_annotation = self.parse_ts_type_annotation();
         if let Some(type_annotation) = &type_annotation {
             Self::extend_binding_pattern_span_end(type_annotation.span.end, &mut kind);
         } else if optional {
             Self::extend_binding_pattern_span_end(self.prev_token_end, &mut kind);
         }
-        Ok(self.ast.binding_pattern(kind, type_annotation, optional))
+        self.ast.binding_pattern(kind, type_annotation, optional)
     }
 
-    pub(crate) fn parse_binding_pattern_kind(&mut self) -> Result<BindingPatternKind<'a>> {
+    pub(crate) fn parse_binding_pattern_kind(&mut self) -> BindingPatternKind<'a> {
         match self.cur_kind() {
             Kind::LCurly => self.parse_object_binding_pattern(),
             Kind::LBrack => self.parse_array_binding_pattern(),
@@ -37,61 +33,61 @@ impl<'a> ParserImpl<'a> {
         }
     }
 
-    fn parse_binding_pattern_identifier(&mut self) -> Result<BindingPatternKind<'a>> {
-        let ident = self.parse_binding_identifier()?;
-        Ok(BindingPatternKind::BindingIdentifier(self.alloc(ident)))
+    fn parse_binding_pattern_identifier(&mut self) -> BindingPatternKind<'a> {
+        let ident = self.parse_binding_identifier();
+        BindingPatternKind::BindingIdentifier(self.alloc(ident))
     }
 
     /// Section 14.3.3 Object Binding Pattern
-    fn parse_object_binding_pattern(&mut self) -> Result<BindingPatternKind<'a>> {
+    fn parse_object_binding_pattern(&mut self) -> BindingPatternKind<'a> {
         let span = self.start_span();
-        self.expect(Kind::LCurly)?;
+        self.expect(Kind::LCurly);
         let (list, rest) = self.parse_delimited_list_with_rest(
             Kind::RCurly,
             Self::parse_binding_property,
             Self::parse_rest_binding,
-        )?;
+        );
         if let Some(rest) = &rest {
             if !matches!(&rest.argument.kind, BindingPatternKind::BindingIdentifier(_)) {
                 let error = diagnostics::invalid_binding_rest_element(rest.argument.span());
-                return Err(self.set_fatal_error(error));
+                return self.fatal_error(error);
             }
         }
-        self.expect(Kind::RCurly)?;
-        Ok(self.ast.binding_pattern_kind_object_pattern(
+        self.expect(Kind::RCurly);
+        self.ast.binding_pattern_kind_object_pattern(
             self.end_span(span),
             list,
             rest.map(|r| self.alloc(r)),
-        ))
+        )
     }
 
     /// Section 14.3.3 Array Binding Pattern
-    fn parse_array_binding_pattern(&mut self) -> Result<BindingPatternKind<'a>> {
+    fn parse_array_binding_pattern(&mut self) -> BindingPatternKind<'a> {
         let span = self.start_span();
-        self.expect(Kind::LBrack)?;
+        self.expect(Kind::LBrack);
         let (list, rest) = self.parse_delimited_list_with_rest(
             Kind::RBrack,
             Self::parse_array_binding_element,
             Self::parse_rest_binding,
-        )?;
-        self.expect(Kind::RBrack)?;
-        Ok(self.ast.binding_pattern_kind_array_pattern(
+        );
+        self.expect(Kind::RBrack);
+        self.ast.binding_pattern_kind_array_pattern(
             self.end_span(span),
             list,
             rest.map(|r| self.alloc(r)),
-        ))
+        )
     }
 
-    fn parse_array_binding_element(&mut self) -> Result<Option<BindingPattern<'a>>> {
+    fn parse_array_binding_element(&mut self) -> Option<BindingPattern<'a>> {
         if self.at(Kind::Comma) {
-            Ok(None)
+            None
         } else {
-            self.parse_binding_pattern_with_initializer().map(Some)
+            Some(self.parse_binding_pattern_with_initializer())
         }
     }
 
-    fn parse_rest_binding(&mut self) -> Result<BindingRestElement<'a>> {
-        let elem = self.parse_rest_element()?;
+    fn parse_rest_binding(&mut self) -> BindingRestElement<'a> {
+        let elem = self.parse_rest_element();
         if let Some(ty) = &elem.argument.type_annotation {
             self.error(diagnostics::rest_element_property_name(ty.span));
         }
@@ -105,16 +101,16 @@ impl<'a> ParserImpl<'a> {
                 self.error(diagnostics::binding_rest_element_last(elem.span));
             }
         }
-        Ok(elem)
+        elem
     }
 
     /// Section 14.3.3 Binding Rest Property
-    pub(super) fn parse_rest_element(&mut self) -> Result<BindingRestElement<'a>> {
+    pub(super) fn parse_rest_element(&mut self) -> BindingRestElement<'a> {
         let span = self.start_span();
         self.bump_any(); // advance `...`
         let init_span = self.start_span();
 
-        let kind = self.parse_binding_pattern_kind()?;
+        let kind = self.parse_binding_pattern_kind();
         // Rest element does not allow `?`, checked in checker/typescript.rs
         if self.at(Kind::Question) && self.is_ts {
             let span = self.cur_token().span();
@@ -122,30 +118,30 @@ impl<'a> ParserImpl<'a> {
             self.error(diagnostics::a_rest_parameter_cannot_be_optional(span));
         }
         // The span is not extended to its type_annotation
-        let type_annotation = self.parse_ts_type_annotation()?;
+        let type_annotation = self.parse_ts_type_annotation();
         let pattern = self.ast.binding_pattern(kind, type_annotation, false);
 
         // Rest element does not allow `= initializer`
         // function foo([...x = []]) { }
         //                    ^^^^ A rest element cannot have an initializer
         let argument = self
-            .context(Context::In, Context::empty(), |p| p.parse_initializer(init_span, pattern))?;
+            .context(Context::In, Context::empty(), |p| p.parse_initializer(init_span, pattern));
         if let BindingPatternKind::AssignmentPattern(pat) = &argument.kind {
             self.error(diagnostics::a_rest_element_cannot_have_an_initializer(pat.span));
         }
 
-        Ok(self.ast.binding_rest_element(self.end_span(span), argument))
+        self.ast.binding_rest_element(self.end_span(span), argument)
     }
 
     /// `BindingProperty`[Yield, Await] :
     ///     `SingleNameBinding`[?Yield, ?Await]
     ///     `PropertyName`[?Yield, ?Await] : `BindingElement`[?Yield, ?Await]
-    pub(super) fn parse_binding_property(&mut self) -> Result<BindingProperty<'a>> {
+    pub(super) fn parse_binding_property(&mut self) -> BindingProperty<'a> {
         let span = self.start_span();
 
         let mut shorthand = false;
         let is_binding_identifier = self.cur_kind().is_binding_identifier();
-        let (key, computed) = self.parse_property_name()?;
+        let (key, computed) = self.parse_property_name();
 
         let value = if is_binding_identifier && !self.at(Kind::Colon) {
             // let { a = b } = c
@@ -156,36 +152,32 @@ impl<'a> ParserImpl<'a> {
                 let identifier =
                     self.ast.binding_pattern_kind_binding_identifier(ident.span, ident.name);
                 let left = self.ast.binding_pattern(identifier, NONE, false);
-                self.context(Context::In, Context::empty(), |p| p.parse_initializer(span, left))?
+                self.context(Context::In, Context::empty(), |p| p.parse_initializer(span, left))
             } else {
-                return Err(self.unexpected());
+                return self.unexpected();
             }
         } else {
             // let { a: b } = c
             //       ^ IdentifierReference
-            self.expect(Kind::Colon)?;
-            self.parse_binding_pattern_with_initializer()?
+            self.expect(Kind::Colon);
+            self.parse_binding_pattern_with_initializer()
         };
 
-        Ok(self.ast.binding_property(self.end_span(span), key, value, shorthand, computed))
+        self.ast.binding_property(self.end_span(span), key, value, shorthand, computed)
     }
 
     /// Initializer[In, Yield, Await] :
     ///   = `AssignmentExpression`[?In, ?Yield, ?Await]
-    fn parse_initializer(
-        &mut self,
-        span: u32,
-        left: BindingPattern<'a>,
-    ) -> Result<BindingPattern<'a>> {
+    fn parse_initializer(&mut self, span: u32, left: BindingPattern<'a>) -> BindingPattern<'a> {
         if self.eat(Kind::Eq) {
-            let expr = self.parse_assignment_expression_or_higher()?;
-            Ok(self.ast.binding_pattern(
+            let expr = self.parse_assignment_expression_or_higher();
+            self.ast.binding_pattern(
                 self.ast.binding_pattern_kind_assignment_pattern(self.end_span(span), left, expr),
                 NONE,
                 false,
-            ))
+            )
         } else {
-            Ok(left)
+            left
         }
     }