… crates/oxc_parser/src/js/class.rs
… crates/oxc_parser/src/js/declaration.rs
… crates/oxc_parser/src/js/expression.rs
… crates/oxc_parser/src/js/function.rs
… crates/oxc_parser/src/js/grammar.rs
… crates/oxc_parser/src/js/module.rs
… crates/oxc_parser/src/js/object.rs
… crates/oxc_parser/src/js/statement.rs
… crates/oxc_parser/src/jsx/mod.rs
… crates/oxc_parser/src/lib.rs
… crates/oxc_parser/src/modifiers.rs
… crates/oxc_parser/src/ts/statement.rs
… crates/oxc_parser/src/ts/types.rs
… tasks/coverage/snapshots/parser_babel.snap
… tasks/coverage/snapshots/parser_test262.snap
… tasks/coverage/snapshots/parser_typescript.snap
… tasks/coverage/snapshots/semantic_babel.snap